adds Retrodock class to support single retrodock jobs from commandline ; reworks RetrodockJob and  JobScheduler job submission functionality

+++ pydock3/jobs.py
@@ -2,6 +2,7 @@ import logging
 import os
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, fields
+from enum import Enum
 
 from pydock3.files import Dir, File, IndockFile
 from pydock3.blastermaster.util import DockFiles
@@ -22,10 +23,18 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 
+#
+class JobSubmissionResult(Enum):
+    SUCCESS = 1
+    FAILED = 2
+    SKIPPED_BECAUSE_ALREADY_COMPLETE = 3
+    SKIPPED_BECAUSE_STILL_RUNNING = 4
+
+
 @dataclass
 class DockingJob(ABC):
     @abstractmethod
-    def run(self):
+    def submit(self):
         raise NotImplementedError
 
 
@@ -51,10 +60,20 @@ class RetrodockJob(ABC):
     def __post_init__(self):
         self._is_complete = False
 
-    def run(self, job_timeout_minutes=None, skip_if_complete=True):
+    def submit(self, job_timeout_minutes=None, skip_if_complete=True):
+        """
+        if job submission is skipped, returns (JobSubmissionResult, None)
+        if job submission is not skipped, returns (JobSubmissionResult, subprocess.CompletedProcess)
+        """
+
+        #
+        if self.is_running:
+            return JobSubmissionResult.SKIPPED_BECAUSE_STILL_RUNNING, None
+
+        #
         if skip_if_complete:
             if self.is_complete:
-                return
+                return JobSubmissionResult.SKIPPED_BECAUSE_ALREADY_COMPLETE, None
 
         # reset output dir
         self.output_dir.delete()
@@ -89,8 +108,8 @@ class RetrodockJob(ABC):
             "ONLY_EXPORT_MOL2_FOR_TASK_1": "true",
         }
 
-        # run job
-        proc = self.job_scheduler.run(
+        # submit job
+        proc = self.job_scheduler.submit(
             job_name=self.name,
             script_path=DOCK_RUN_SCRIPT_PATH,
             env_vars_dict=env_vars_dict,
@@ -100,7 +119,7 @@ class RetrodockJob(ABC):
         )
         self.num_attempts += 1
 
-        return proc
+        return JobSubmissionResult, proc
 
     @property
     def is_running(self):
